deepboard/requirements.txt
python-fasthtml
fh-plotly
plotly
pandas
markupsafe
